Added the item viewer
Added a little display that shows what item is inside of an inventory when it is hovered. Also moved the machine preview script to the machine preview object (and tidied it up a bit), removed some unnecessary references, and started using Action<T> for my events where it is applicable.