Making progress on grouping machines
Grouping machines is taking a while to figure out, but it's coming along well. I've had to mess with and make some decisions about how I want to structure recipes. Most of the remaining work should just be implementing the actual process of calculating the recipe when creating a machine group.